test(sidekick): fix `config` package coverage
Move tests around so they get counted towards coverage.

Add test for one function.
