+++ internal/sidekick/internal/config/config.go
@@ -36,6 +36,7 @@ const (
 	defaultGitHub    = "https://github.com"
 	repo             = "googleapis/googleapis"
 	branch           = "master"
+	configName       = ".sidekick.toml"
 )
 
 // DocumentationOverride describes overrides for the documentation of a single element.
@@ -78,7 +79,7 @@ type GeneralConfig struct {
 // Where applicable, overrides the top level (or default) configuration values with the ones passed in the command line.
 // Returns the merged configuration, or an error if the top level configuration is invalid.
 func LoadConfig(language string, source, codec map[string]string) (*Config, error) {
-	rootConfig, err := LoadRootConfig(".sidekick.toml")
+	rootConfig, err := LoadRootConfig(configName)
 	if err != nil {
 		return nil, err
 	}
